refactor(entities)!: move serializers from apis_entities to entities
BREAKING-CHANGE: The `E21_PersonCidocSerializer` and the
`E53_PlaceCidocSerializer` are beim moved to `apis_core.entities`
